+++ demos/django-react-native-todolist/CHANGELOG.md
@@ -1,5 +1,15 @@
 # django-react-native-todolist
 
+## 1.1.4
+
+### Patch Changes
+
+- Updated dependencies [62e43aa]
+- Updated dependencies [6b01811]
+  - @powersync/react-native@1.6.0
+  - @powersync/common@1.9.0
+  - @powersync/react@1.3.4
+
 ## 1.1.3
 
 ### Patch Changes