+++ demos/example-nextjs/CHANGELOG.md
@@ -1,5 +1,14 @@
 # example-nextjs
 
+## 0.1.4
+
+### Patch Changes
+
+- Updated dependencies [62e43aa]
+- Updated dependencies [6b01811]
+  - @powersync/web@0.7.0
+  - @powersync/react@1.3.4
+
 ## 0.1.3
 
 ### Patch Changes