+++ demos/example-vite/CHANGELOG.md
@@ -1,5 +1,13 @@
 # example-vite
 
+## 0.0.22
+
+### Patch Changes
+
+- Updated dependencies [62e43aa]
+- Updated dependencies [6b01811]
+  - @powersync/web@0.7.0
+
 ## 0.0.21
 
 ### Patch Changes