+++ packages/common/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @powersync/common
 
+## 1.9.0
+
+### Minor Changes
+
+- 62e43aa: Improved import and usage of BSON library.
+
 ## 1.8.1
 
 ### Patch Changes